python/requirements.txt
pyzmq==25.1.2
protobuf==4.25.1
numpy==1.26.3
pandas==2.1.4
polars==1.38.1
connectorx==0.4.5
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pyyaml==6.0.1
python-dotenv==1.2.1
prometheus-client==0.19.0
